requirements.txt
numpy==2.2.4
Pillow==11.1.0
protobuf==6.30.1
python-dotenv==1.0.1
SQLAlchemy==2.0.39
streamlit==1.43.2
tensorflow==2.19.0
